# lib/api-zod/src/generated/types/quizLeadInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizLeadInputAnswers } from "./quizLeadInputAnswers";

export interface QuizLeadInput {
  email?: string;
  answers: QuizLeadInputAnswers;
}